Adding some helper methods
