+++ src/components/circle.rs
@@ -1,4 +1,4 @@
-﻿use crate::components::Position;
+﻿use crate::components::position::Position;
 use crate::MaybeSignalOption;
 use leaflet::CircleOptions;
 use leptos::*;
